+++ cmake/ParcoachCommonOptions.cmake
@@ -1,4 +1,5 @@
 option(PARCOACH_ENABLE_INSTRUMENTATION "Build PARCOACH instrumentation module." ON)
+option(PARCOACH_ENABLE_FORTRAN "Enable Fortran's instrumentation support through LLVM's Flang." OFF)
 option(PARCOACH_ENABLE_MPI "Build PARCOACH with MPI support." ON)
 option(PARCOACH_ENABLE_OPENMP "Build PARCOACH with OpenMP support." ON)
 option(PARCOACH_ENABLE_CUDA "Build PARCOACH with Cuda support." OFF)