+++ src/rma/RMAStatisticsAnalysis.cpp
@@ -1,67 +1,36 @@
-#include "parcoach/LocalConcurrencyDetectionPass.h"
+#include "parcoach/RMAPasses.h"
 #include "llvm/IR/InstIterator.h"
 
 using namespace llvm;
-namespace parcoach {
+namespace parcoach::rma {
 
 namespace {
-// Count MPI operations in Fortran
-void CountMPIfuncFORTRAN(RMAStatisticsAnalysis::Result &Res, Instruction &I) {
-  if (I.getOperand(0)->getName() == ("mpi_put_")) {
-    // GreenErr() << "PARCOACH DEBUG: Found Put\n ";
-    // DEBUG INFO: I.print(errs(),nullptr);
-    Res.Put++;
-  } else if (I.getOperand(0)->getName() == ("mpi_get_")) {
-    Res.Get++;
-  } else if (I.getOperand(0)->getName() == ("mpi_accumulate_")) {
-    Res.Acc++;
-  } else if (I.getOperand(0)->getName() == ("mpi_win_create_")) {
-    Res.Win++;
-  } else if (I.getOperand(0)->getName() == ("mpi_win_free_")) {
-    Res.Free++;
-  } else if (I.getOperand(0)->getName() == ("mpi_win_fence_")) {
-    Res.Fence++;
-  } else if (I.getOperand(0)->getName() == ("mpi_win_unlock_all_")) {
-    Res.Unlockall++;
-  } else if (I.getOperand(0)->getName() == ("mpi_win_unlock_")) {
-    Res.Unlock++;
-  } else if (I.getOperand(0)->getName() == ("mpi_win_lock_")) {
-    Res.Lock++;
-  } else if (I.getOperand(0)->getName() == ("mpi_win_lock_all_")) {
-    Res.Lockall++;
-  } else if (I.getOperand(0)->getName() == ("mpi_win_flush_")) {
-    Res.Flush++;
-  } else if (I.getOperand(0)->getName() == ("mpi_barrier_")) {
-    Res.Barrier++;
-  }
-}
-
-void CountMPIfuncC(RMAStatisticsAnalysis::Result &Res, Instruction &I,
+void CountMPIfuncs(RMAStatisticsAnalysis::Result &Res, Instruction &I,
                    StringRef Name) {
 
-  if (Name == "MPI_Get") {
+  if (Name == "MPI_Get" || Name == "mpi_get_") {
     Res.Get++;
-  } else if (Name == "MPI_Put") {
+  } else if (Name == "MPI_Put" || Name == "mpi_put_") {
     Res.Put++;
-  } else if (Name == "MPI_Win_create") {
+  } else if (Name == "MPI_Win_create" || Name == "mpi_win_create_") {
     Res.Win++;
-  } else if (Name == "MPI_Accumulate") {
+  } else if (Name == "MPI_Accumulate" || Name == "mpi_accumulate_") {
     Res.Acc++;
-  } else if (Name == "MPI_Win_fence") {
+  } else if (Name == "MPI_Win_fence" || Name == "mpi_win_fence_") {
     Res.Fence++;
-  } else if (Name == "MPI_Win_flush") {
+  } else if (Name == "MPI_Win_flush" || Name == "mpi_win_flush_") {
     Res.Flush++;
-  } else if (Name == "MPI_Win_lock") {
+  } else if (Name == "MPI_Win_lock" || Name == "mpi_win_lock_") {
     Res.Lock++;
-  } else if (Name == "MPI_Win_unlock") {
+  } else if (Name == "MPI_Win_unlock" || Name == "mpi_win_unlock_") {
     Res.Unlock++;
-  } else if (Name == "MPI_Win_unlock_all") {
+  } else if (Name == "MPI_Win_unlock_all" || Name == "mpi_win_unlock_all_") {
     Res.Unlockall++;
-  } else if (Name == "MPI_Win_lock_all") {
+  } else if (Name == "MPI_Win_lock_all" || Name == "mpi_win_lock_all_") {
     Res.Lockall++;
-  } else if (Name == "MPI_Win_free") {
+  } else if (Name == "MPI_Win_free" || Name == "mpi_win_free_") {
     Res.Free++;
-  } else if (Name == "MPI_Barrier") {
+  } else if (Name == "MPI_Barrier" || Name == "mpi_barrier_") {
     Res.Barrier++;
   }
 }
@@ -73,19 +42,14 @@ RMAStatisticsAnalysis::Statistics
 RMAStatisticsAnalysis::run(Function &F, FunctionAnalysisManager &) {
   TimeTraceScope TTS("RMAStatisticsAnalysis");
   Statistics Res{};
-
   for (Instruction &I : instructions(F)) {
     DebugLoc dbg = I.getDebugLoc(); // get debug infos
-    if (I.getOpcode() == Instruction::BitCast) {
-      if (I.getOperand(0)->getName().startswith("mpi_")) {
-        Res.Mpi++;
-        CountMPIfuncFORTRAN(Res, I);
-      }
-    } else if (CallBase *cb = dyn_cast<CallBase>(&I)) {
+    if (CallBase *cb = dyn_cast<CallBase>(&I)) {
       if (Function *calledFunction = cb->getCalledFunction()) {
-        if (calledFunction->getName().startswith("MPI_")) {
+        StringRef FName = calledFunction->getName();
+        if (FName.startswith("MPI_") || FName.startswith("mpi_")) {
           Res.Mpi++;
-          CountMPIfuncC(Res, I, calledFunction->getName());
+          CountMPIfuncs(Res, I, calledFunction->getName());
         }
       }
     }
@@ -102,4 +66,4 @@ size_t RMAStatisticsAnalysis::Statistics::getTotalOneSided() const {
   return Put + Get + Acc;
 }
 
-} // namespace parcoach
+} // namespace parcoach::rma